backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
anthropic==0.40.0
openpyxl==3.1.5
weasyprint==62.3
python-dotenv==1.0.1
pydantic==2.10.3
